bessctl: Add GRAPHEASY_OPTS to show/monitor pipeline.
Similar to TCPDUMP_OPTS, this lets the user provide additional options
to the graph-easy invocation (or even pipe the graph-easy output to other
programs).

I find this change useful because sometimes the graph is too complicated to
be shown on the terminal.  With this commit

```
localhost:10514 $ show pipeline --as dot | dot -Tsvg -o graph.svg
```

outputs the graph to a file on disk.